Updating 9.4 (#38)
* Adjusted Distr Command to add Icons to the list,
Hopefully it now uses the localizer to use the Localized block names on output
Still requires information about the full space and save output

* Fix for trees

Fix for trees not deleted and not clearing  #19
Also trees now correcly can be exported and imported
Changed version number

* Undo command usability

Optional argument for `/undo` allow to perform multiple undone with one command #20

* Distr command

Reworked, changed output formatting.
Now takes optional arguments:
brief (b) - short output (default if no args)
detail (d) - detailed output (can allso understand detailed, details)

* Author info

Added author information handling for import/export related to #21

* Remove whitespace from project name and output dll, folder for schematics

Schematic folder moved to constant place (Storage/Schematics) #10
Rename project and generated DLL removing whitespace

* Fix for inventory

Fixed paste broke #27 on object with inventory when trying add items with quantity more than max stack size. Show error to player and try to fix quantity.
Also fix for suddenly renamed expand command

* Fix for undo, locking, file info

Fix for #28 fixed clearance for objects not recording, fixed objects overwrite with blocks - place blocks first then objects in undo
Lock command execution when previous command still running
Introduce file info for #21
A little bit of refactoring/improvements

* Rework and critical fix

Fully reworked selection handling. Now uses WorldRegion
Fix the issue (WorldRegion loops don't have that problem) with possible infinity loop when SVP returns one axis as negative

* Fix for stockpiles

Fixed that full stockpiles was not be detected and copied or deleted and so on. Not the best solution but there no other way atm

* Distr file output and foreach optimize

Distr selection info and file output #18
Optimize foreach loops

* Import imrovements

Build and show list of schematic files #21
Cache file list and recache for speed up. loading list at server start
+some fixes

* Cleanup

Clean debugging messages
Changed date in blueptint list printing to modified date as overwrite file not change creation date

* Cut command and changes

Introduce and switch to providing selection for Execute method from base class allowing to perform command invocation with custom selection range
Added /cut command #31

* Stack offset param

New parameter for /stack command #32 give negative number like -1 allows to chain rooms without creating double walls between them
Updated README

* Claim selection, fixes, rename of contract coomand

Added /selclaim, /addclaim, /expclaim commands #23
Rename /contract to /reduce #36
Fixes for /stack, /move can get out of world bounds
Fix some commands still able replace impenetrable stone
Fix for Y coords swap - now just limited

* add Wand Tool icon

* Temp Icon and tool for WE Wand

+++ WorldEdit/WorldEditBlockManager.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Eco.Core.IoC;
+using Eco.Core.Utils;
+using Eco.Gameplay.Blocks;
 using Eco.Gameplay.Components;
 using Eco.Gameplay.Objects;
 using Eco.Gameplay.Plants;
-using Eco.Gameplay.Players;
 using Eco.Mods.WorldEdit.Model;
-using Eco.Shared.Localization;
 using Eco.Shared.Math;
 using Eco.Shared.Utils;
 using Eco.Simulation;
@@ -25,26 +26,27 @@ namespace Eco.Mods.WorldEdit
 		public static void SetBlock(Type type, Vector3i position)
 		{
 			ClearPosition(position);
-			RestoreBlock(type, position);
+			if (type != typeof(EmptyBlock)) RestoreBlock(type, position);
 		}
-		public static void RestoreBlockOffset(WorldEditBlock block, Vector3i offsetPos, Player player)
+		public static void RestoreBlockOffset(WorldEditBlock block, Vector3i offsetPos, UserSession session)
 		{
-			RestoreBlock(block, ApplyOffset(block.Position, offsetPos), player);
+			RestoreBlock(block, ApplyOffset(block.Position, offsetPos), session);
 		}
-		public static void RestoreBlock(WorldEditBlock block, Vector3i position, Player player)
+		public static void RestoreBlock(WorldEditBlock block, Vector3i position, UserSession session)
 		{
+			if (IsImpenetrable(position)) return;
 			ClearPosition(position);
-			if (block.BlockType.Equals(typeof(EmptyBlock)))
+			if (block.IsEmptyBlock())
 			{
 				RestoreEmptyBlock(position);
 			}
-			else if (block.BlockType.DerivesFrom<PlantBlock>())
+			else if (block.IsPlantBlock())
 			{
 				RestorePlantBlock(block.BlockType, position, block.BlockData);
 			}
-			else if (block.BlockType.DerivesFrom<WorldObjectBlock>())
+			else if (block.IsWorldObjectBlock())
 			{
-				RestoreWorldObjectBlock(block.BlockType, position, block.BlockData, player);
+				RestoreWorldObjectBlock(block.BlockType, position, block.BlockData, session);
 			}
 			else
 			{
@@ -54,28 +56,31 @@ namespace Eco.Mods.WorldEdit
 
 		public static void RestoreEmptyBlock(Vector3i position)
 		{
+			if (IsImpenetrable(position)) return;
 			World.DeleteBlock(position);
 		}
 
 		public static void RestoreBlock(Type type, Vector3i position)
 		{
+			if (IsImpenetrable(position)) return;
 			World.SetBlock(type, position);
 		}
 
-		public static void RestoreWorldObjectBlock(Type type, Vector3i position, IWorldEditBlockData blockData, Player player)
+		public static void RestoreWorldObjectBlock(Type type, Vector3i position, IWorldEditBlockData blockData, UserSession session)
 		{
+			if (blockData == null) { return; }
 			WorldEditWorldObjectBlockData worldObjectBlockData = (WorldEditWorldObjectBlockData)blockData;
-			ClearWorldObjectPlace(worldObjectBlockData.WorldObjectType, position, worldObjectBlockData.Rotation);
+			ClearWorldObjectPlace(worldObjectBlockData.WorldObjectType, position, worldObjectBlockData.Rotation, session);
 
 			WorldObject worldObject = null;
-			try { worldObject = WorldObjectManager.ForceAdd(worldObjectBlockData.WorldObjectType, player.User, position, worldObjectBlockData.Rotation, true); }
+			try { worldObject = WorldObjectManager.ForceAdd(worldObjectBlockData.WorldObjectType, session.User, position, worldObjectBlockData.Rotation, true); }
 			catch (Exception e)
 			{
-				Log.WriteLine(new LocString(e.ToString()));
+				Log.WriteException(e);
 			}
 			if (worldObject == null)
 			{
-				Log.WriteLine(Localizer.Do($"Unable spawn WorldObject {worldObjectBlockData.WorldObjectType} at {position}"));
+				Log.WriteErrorLineLoc($"Unable spawn WorldObject {worldObjectBlockData.WorldObjectType} at {position}");
 				return;
 			}
 			if (worldObject.HasComponent<StorageComponent>() && worldObjectBlockData.Components.ContainsKey(typeof(StorageComponent)))
@@ -96,7 +101,12 @@ namespace Eco.Mods.WorldEdit
 				foreach (InventoryStack stack in inventoryStacks)
 				{
 					if (stack.ItemType == null) continue;
-					storageComponent.Inventory.AddItems(stack.ItemType, stack.Quantity);
+					Result result = storageComponent.Inventory.TryAddItems(stack.ItemType, stack.Quantity);
+					if (result.Failed)
+					{
+						session.Player.ErrorLocStr(result.Message.Trim());
+						try { storageComponent.Inventory.AddItems(stack.GetItemStack()); } catch (InvalidOperationException) { /*Already show error to user*/ }
+					}
 				}
 			}
 			if (worldObject.HasComponent<CustomTextComponent>() && worldObjectBlockData.Components.ContainsKey(typeof(CustomTextComponent)))
@@ -108,9 +118,17 @@ namespace Eco.Mods.WorldEdit
 
 		public static void RestorePlantBlock(Type type, Vector3i position, IWorldEditBlockData blockData)
 		{
+			if (blockData == null) { return; }
 			WorldEditPlantBlockData plantBlockData = (WorldEditPlantBlockData)blockData;
-			PlantSpecies plantSpecies = EcoSim.AllSpecies.OfType<PlantSpecies>().First(species => species.BlockType == type);
-			Plant plant = EcoSim.PlantSim.SpawnPlant(plantSpecies, position);
+			PlantSpecies plantSpecies = null;
+			try { plantSpecies = EcoSim.AllSpecies.OfType<PlantSpecies>().First(species => species.GetType() == plantBlockData.PlantType); }
+			catch (InvalidOperationException)
+			{
+				//TODO: Temporary support for the old serialized format! Should be done with migration!
+				plantSpecies = EcoSim.AllSpecies.OfType<PlantSpecies>().First(species => species.Name == plantBlockData.PlantType.Name);
+			}
+			if (plantSpecies == null) return;
+			Plant plant = EcoSim.PlantSim.SpawnPlant(plantSpecies, position, true);
 			plant.YieldPercent = plantBlockData.YieldPercent;
 			plant.Dead = plantBlockData.Dead;
 			plant.DeadType = plantBlockData.DeadType;
@@ -131,14 +149,27 @@ namespace Eco.Mods.WorldEdit
 					worldObjectBlock.WorldObjectHandle.Object.Destroy();
 					break;
 				case ImpenetrableStoneBlock _:
+					return;
+				case PlantBlock plantBlock:
+				case TreeBlock treeBlock:
+					Plant plant = EcoSim.PlantSim.GetPlant(position);
+					if (plant != null) { EcoSim.PlantSim.DestroyPlant(plant, DeathType.DivineIntervention, true); }
 					break;
 				default:
-					World.DeleteBlock(position);
+					if (BlockContainerManager.Obj.IsBlockContained(position))
+					{
+						WorldObject worldObject = ServiceHolder<IWorldObjectManager>.Obj.All.Where(x => x.Position3i.Equals(position)).FirstOrDefault();
+						if (worldObject != null) worldObject.Destroy();
+					}
+					else
+					{
+						World.DeleteBlock(position);
+					}
 					break;
 			}
 		}
 
-		private static void ClearWorldObjectPlace(Type worldObjectType, Vector3i position, Quaternion rotation)
+		private static void ClearWorldObjectPlace(Type worldObjectType, Vector3i position, Quaternion rotation, UserSession session)
 		{
 			List<BlockOccupancy> blockOccupancies = WorldObject.GetOccupancy(worldObjectType);
 			if (!blockOccupancies.Any(x => x.BlockType != null)) return;
@@ -147,6 +178,7 @@ namespace Eco.Mods.WorldEdit
 				if (blockOccupancy.BlockType != null)
 				{
 					Vector3i worldPos = position + rotation.RotateVector(blockOccupancy.Offset).XYZi;
+					if (!session.ExecutingCommand.PerformingUndo) session.ExecutingCommand.AddBlockChangedEntry(worldPos); //Do not record changes when doing undo
 					ClearPosition(worldPos);
 				}
 			}
